+++ src/pytest_shared_session_scope/__init__.py
@@ -3,10 +3,12 @@
 from pytest_shared_session_scope.types import (
     Store as Store,
     CleanupToken as CleanupToken,
+    SetupToken as SetupToken,
     StoreValueNotExists as StoreValueNotExists,
 )
 
 from pytest_shared_session_scope.fixtures import (
     shared_session_scope_fixture as shared_session_scope_fixture,
     shared_session_scope_json as shared_session_scope_json,
+    shared_session_scope_pickle as shared_session_scope_pickle,
 )