+++ src/pytest_shared_session_scope/types.py
@@ -1,14 +1,12 @@
 """Common types used in the package."""
 
-from enum import Enum
+from enum import Enum, auto
 from contextlib import AbstractContextManager
 from typing import Any, Generic, Protocol, TypeVar
 
 
 class StoreValueNotExists(Exception):
     """Raised when a value is not found in the storage."""
-
-    ...
 
 
 _StoreType = TypeVar("_StoreType")
@@ -39,7 +37,13 @@ class Store(Protocol, Generic[_StoreType]):
         ...
 
 
-class CleanupToken(str, Enum):
-    """Token that is send back to the fixture after last yield."""
+class SetupToken(Enum):
+    """Token that is send back to the fixture in first yield."""
 
-    LAST = "last"
+    FIRST = auto()
+
+
+class CleanupToken(Enum):
+    """Token that is send back to the fixture in last yield."""
+
+    LAST = auto()