+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 ## [Unreleased]
+
+- Add `shared_session_scope_pickle` that uses `pickle` for de/serialization.
+Make fixture that does cleanup also save missing test (which is then an empty set). Should make no runtime in supported use cases.
+- Make fixture that does cleanup also save missing test (which is then an empty set). Should make no runtime in supported use cases.
+
+## [0.4.0]
+- Changed `CleanupToken` to be just `Enum` instead of `str, Enum`.
+- BREAKING: The first yield now returns a `SetupToken` instead of `None` to signal that a worker should calculate the value.
+
+## [0.3.0]
 - BREAKING: Rename `shared_json_session_scope` to `shared_session_scope_json`
 
 ## [0.3.0]